requirements.txt
djongo==1.3.6
djangorestframework
django==4.1.9
pymongo==3.12.3
django-cors-headers
